clean up authenticated_user_spec, rename to authenticated_traveler, move specs for other features to other specs
